+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 ### Added
+- configure.py script, to download, build and install mpir 3.0.0
 
 ### Changed
+- splitted lib/Exceptions.ark into lib/Exceptions.ark and lib/Either.ark
 
 ### Removed
 - `hastype` keyword because I never had to implement compile time typechecking, so it's not useful